requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
pymongo==4.7.2
bcrypt==4.1.3
python-dotenv==1.0.1
spacy==3.7.4
nltk==3.8.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
scikit-learn==1.5.0
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
streamlit==1.35.0
altair==5.3.0
pandas==2.2.2
numpy==1.26.4
transformers==4.46.3
torch==2.5.1
tf-keras
google-generativeai